+++ src/components/StyledText.tsx
