+++ src/components/Themed.tsx
@@ -6,8 +6,8 @@
 import * as React from 'react';
 import { Text as DefaultText, View as DefaultView } from 'react-native';
 
-import Colors from '../constants/Colors';
-import useColorScheme from '../hooks/useColorScheme';
+import Colors from '../../constants/Colors';
+import useColorScheme from '../../hooks/useColorScheme';
 
 export function useThemeColor(
   props: { light?: string; dark?: string },
