+++ src/screens/NotFoundScreen.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { StyleSheet, TouchableOpacity } from 'react-native';
 import { Text, View } from '../components/Themed';
 
-import { RootStackScreenProps } from '../types';
+import { RootStackScreenProps } from '../../types';
 
 export default function NotFoundScreen({ navigation }: RootStackScreenProps<'NotFound'>) {
   return (
